Fix Problem w/ Dropping "Title" Var
Before had an issue with year 2003 where the "all" file didn't have a title variable and was breaking the code so I just used "cap". But now fixed it so that we don't use cap and instead put the 2003 all file in with the priority so that the "title" variable isn't specified.